feat: implemented user email existing check during registration

+++ src/main/java/com/e_commerce/e_commerce/controller/HomeController.java
@@ -144,23 +144,31 @@ public class HomeController {
     public String saveUser(@ModelAttribute UserDtls user, @RequestParam("img") MultipartFile file, HttpSession session)
             throws IOException {
 
-        String imageName = file.isEmpty() ? "default.jpg" : file.getOriginalFilename();
-        user.setProfileImage(imageName);
-        UserDtls saveUser = userService.saveUser(user);
+        Boolean existsEmail = userService.existsEmail(user.getEmail());
+        if(existsEmail)
+        {
+            session.setAttribute("errorMsg", "Email already exists");
 
-        if (!ObjectUtils.isEmpty(saveUser)) {
-            if (!file.isEmpty()) {
-                File saveFile = new ClassPathResource("static/img").getFile();
+        }else {
+            String imageName = file.isEmpty() ? "default.jpg" : file.getOriginalFilename();
+            user.setProfileImage(imageName);
+            UserDtls saveUser = userService.saveUser(user);
 
-                Path path = Paths.get(saveFile.getAbsolutePath() + File.separator + "profile_img" + File.separator
-                        + file.getOriginalFilename());
+            if (!ObjectUtils.isEmpty(saveUser)) {
+                if (!file.isEmpty()) {
+                    File saveFile = new ClassPathResource("static/img").getFile();
+
+                    Path path = Paths.get(saveFile.getAbsolutePath() + File.separator + "profile_img" + File.separator
+                            + file.getOriginalFilename());
 
 //				System.out.println(path);
-                Files.copy(file.getInputStream(), path, StandardCopyOption.REPLACE_EXISTING);
+                    Files.copy(file.getInputStream(), path, StandardCopyOption.REPLACE_EXISTING);
+                }
+                session.setAttribute("succMsg", "Register successfully");
+            } else {
+                session.setAttribute("errorMsg", "something wrong on server");
             }
-            session.setAttribute("succMsg", "Register successfully");
-        } else {
-            session.setAttribute("errorMsg", "something wrong on server");
+
         }
 
         return "redirect:/register";
